refactor: reorganize unit tests to follow src directory structure
