+++ test/unit/_MISC_optional_require.test.js
