+++ test/unit/error.test.js
@@ -1,0 +1,209 @@
+/* eslint no-empty: ["error", { "allowEmptyCatch": true }] */
+'use strict';
+
+const expect = require('chai').expect;
+const mock = require('../tools/mock');
+const { getSymbolFrom } = require('../tools/utils');
+const { ReplSetFixture } = require('../tools/common');
+const { ns } = require('../../src/utils');
+const { Topology } = require('../../src/sdam/topology');
+const { MongoNetworkError, MongoWriteConcernError } = require('../../src/index');
+const {
+  PoolClosedError: MongoPoolClosedError,
+  WaitQueueTimeoutError: MongoWaitQueueTimeoutError
+} = require('../../src/cmap/errors');
+
+describe('MongoErrors', () => {
+  // import errors as object
+  let errorClasses = Object.fromEntries(
+    Object.entries(require('../../src/index')).filter(([key]) => key.endsWith('Error'))
+  );
+  errorClasses = { ...errorClasses, MongoPoolClosedError, MongoWaitQueueTimeoutError };
+
+  for (const errorName in errorClasses) {
+    describe(errorName, () => {
+      it(`name should be read-only`, () => {
+        // Dynamically create error class with message
+        let error = new errorClasses[errorName]('generated by test');
+        // expect name property to be class name
+        expect(error).to.have.property('name', errorName);
+
+        try {
+          error.name = 'renamed by test';
+        } catch (err) {}
+        expect(error).to.have.property('name', errorName);
+      });
+    });
+  }
+
+  describe('when MongoNetworkError is constructed', () => {
+    it('should only define beforeHandshake symbol if boolean option passed in', function () {
+      const errorWithOptionTrue = new MongoNetworkError('', { beforeHandshake: true });
+      expect(getSymbolFrom(errorWithOptionTrue, 'beforeHandshake', false)).to.be.a('symbol');
+
+      const errorWithOptionFalse = new MongoNetworkError('', { beforeHandshake: false });
+      expect(getSymbolFrom(errorWithOptionFalse, 'beforeHandshake', false)).to.be.a('symbol');
+
+      const errorWithBadOption = new MongoNetworkError('', { beforeHandshake: 'not boolean' });
+      expect(getSymbolFrom(errorWithBadOption, 'beforeHandshake', false)).to.be.an('undefined');
+
+      const errorWithoutOption = new MongoNetworkError('');
+      expect(getSymbolFrom(errorWithoutOption, 'beforeHandshake', false)).to.be.an('undefined');
+    });
+  });
+
+  describe('WriteConcernError', function () {
+    let test;
+    const RAW_USER_WRITE_CONCERN_CMD = {
+      createUser: 'foo2',
+      pwd: 'pwd',
+      roles: ['read'],
+      writeConcern: { w: 'majority', wtimeoutMS: 1 }
+    };
+
+    const RAW_USER_WRITE_CONCERN_ERROR = {
+      ok: 0,
+      errmsg: 'waiting for replication timed out',
+      code: 64,
+      codeName: 'WriteConcernFailed',
+      writeConcernError: {
+        code: 64,
+        codeName: 'WriteConcernFailed',
+        errmsg: 'waiting for replication timed out',
+        errInfo: {
+          wtimeout: true
+        }
+      }
+    };
+
+    const RAW_USER_WRITE_CONCERN_ERROR_INFO = {
+      ok: 0,
+      errmsg: 'waiting for replication timed out',
+      code: 64,
+      codeName: 'WriteConcernFailed',
+      writeConcernError: {
+        code: 64,
+        codeName: 'WriteConcernFailed',
+        errmsg: 'waiting for replication timed out',
+        errInfo: {
+          writeConcern: {
+            w: 2,
+            wtimeout: 0,
+            provenance: 'clientSupplied'
+          }
+        }
+      }
+    };
+
+    before(() => (test = new ReplSetFixture()));
+    afterEach(() => mock.cleanup());
+    beforeEach(() => test.setup());
+
+    function makeAndConnectReplSet(cb) {
+      let invoked = false;
+      const replSet = new Topology(
+        [test.primaryServer.hostAddress(), test.firstSecondaryServer.hostAddress()],
+        { replicaSet: 'rs' }
+      );
+
+      replSet.once('error', err => {
+        if (invoked) {
+          return;
+        }
+        invoked = true;
+        cb(err);
+      });
+
+      replSet.on('connect', () => {
+        if (invoked) {
+          return;
+        }
+
+        invoked = true;
+        cb(undefined, replSet);
+      });
+
+      replSet.connect();
+    }
+
+    it('should expose a user command writeConcern error like a normal WriteConcernError', function (done) {
+      test.primaryServer.setMessageHandler(request => {
+        const doc = request.document;
+        if (doc.ismaster || doc.hello) {
+          setTimeout(() => request.reply(test.primaryStates[0]));
+        } else if (doc.createUser) {
+          setTimeout(() => request.reply(RAW_USER_WRITE_CONCERN_ERROR));
+        }
+      });
+
+      makeAndConnectReplSet((err, topology) => {
+        // cleanup the server before calling done
+        const cleanup = err => topology.close({ force: true }, err2 => done(err || err2));
+
+        if (err) {
+          return cleanup(err);
+        }
+
+        topology.selectServer('primary', (err, server) => {
+          expect(err).to.not.exist;
+
+          server.command(ns('db1'), Object.assign({}, RAW_USER_WRITE_CONCERN_CMD), err => {
+            let _err;
+            try {
+              expect(err).to.be.an.instanceOf(MongoWriteConcernError);
+              expect(err.result).to.exist;
+              expect(err.result).to.have.property('ok', 1);
+              expect(err.result).to.not.have.property('errmsg');
+              expect(err.result).to.not.have.property('code');
+              expect(err.result).to.not.have.property('codeName');
+              expect(err.result).to.have.property('writeConcernError');
+            } catch (e) {
+              _err = e;
+            } finally {
+              cleanup(_err);
+            }
+          });
+        });
+      });
+    });
+
+    it('should propagate writeConcernError.errInfo ', function (done) {
+      test.primaryServer.setMessageHandler(request => {
+        const doc = request.document;
+        if (doc.ismaster || doc.hello) {
+          setTimeout(() => request.reply(test.primaryStates[0]));
+        } else if (doc.createUser) {
+          setTimeout(() => request.reply(RAW_USER_WRITE_CONCERN_ERROR_INFO));
+        }
+      });
+
+      makeAndConnectReplSet((err, topology) => {
+        // cleanup the server before calling done
+        const cleanup = err => topology.close(err2 => done(err || err2));
+
+        if (err) {
+          return cleanup(err);
+        }
+
+        topology.selectServer('primary', (err, server) => {
+          expect(err).to.not.exist;
+
+          server.command(ns('db1'), Object.assign({}, RAW_USER_WRITE_CONCERN_CMD), err => {
+            let _err;
+            try {
+              expect(err).to.be.an.instanceOf(MongoWriteConcernError);
+              expect(err.result).to.exist;
+              expect(err.result.writeConcernError).to.deep.equal(
+                RAW_USER_WRITE_CONCERN_ERROR_INFO.writeConcernError
+              );
+            } catch (e) {
+              _err = e;
+            } finally {
+              cleanup(_err);
+            }
+          });
+        });
+      });
+    });
+  });
+});
